feat: refactor exception handling and add utility classes

+++ src/main/java/com/restaurants/demo/dto/response/DailyReportResponse.java
@@ -13,6 +13,6 @@ import java.math.BigDecimal;
 @Setter
 @AllArgsConstructor
 public class DailyReportResponse {
-    private BigDecimal totalRevenue;
+    private double totalRevenue;
     private Long totalOrders;
 }
